fix: separate batching ops
